Fixes #48 and #58 - Resize plot when window size changes
Added an event_handler to resize event of the canvas which detects
if a function has already been plotted on the canvas, and if yes,
the plot is resized according to the new canvas size. Also stops
the program from plotting when switching between themes, if no function
is provided.